+++ app/state.go
@@ -35,7 +35,7 @@ type AppInfo struct {
 func (a *App) GetAppInfo() AppInfo {
 	return AppInfo{
 		Name:        "Aerion",
-		Version:     "0.1.10",
+		Version:     "0.1.11",
 		Description: "An Open Source Lightweight E-Mail Client",
 		Website:     "https://github.com/hkdb/aerion",
 		License:     "Apache 2.0",